+++ R/sim.R
@@ -177,14 +177,14 @@ sim.words <- function(text1, text2 = NULL, method = res, weight = NULL,
 
 # quick string comparison based on cosine similarity between bigrams
 
-sim.strings <- function(strings1, strings2 = NULL, sep = "", boundary = TRUE) {
+sim.strings <- function(strings1, strings2 = NULL, sep = "", boundary = TRUE, ...) {
 
-	S1 <- splitStrings(strings1, sep = sep, boundary = boundary, simplify = TRUE)
+	S1 <- splitStrings(strings1, sep = sep, boundary = boundary, simplify = TRUE, ...)
 
 	if (is.null(strings2)) {
 		sim <- cosSparse(S1)
 	} else {
-		S2 <- splitStrings(strings2, sep = sep, boundary = boundary, simplify = TRUE)
+		S2 <- splitStrings(strings2, sep = sep, boundary = boundary, simplify = TRUE, ...)
 		M <- jMatrix( rownames(S1), rownames(S2) )
 		sim <- cosSparse( (M$M1*1) %*% S1, (M$M2*1) %*% S2 )
 	}